include: Pass a PROCESS_INFORMATION to winetest_wait_child_process().
And close the handles in that function.

Most callers are simplified. Some callers need helpers to return the whole
PROCESS_INFORMATION instead of just the process handle but are still simplified
overall. In two cases the process handle is used after joining and so a
WaitForSingleObject() is added. In one case, test_Toolhelp(), there is no
PROCESS_INFORMATION and so the function is replaced.

This also avoids GetProcessId(), which does not exist on Windows 98.