dsound: Remove dead lead-in logic.
This is effectively dead code, because if the buffer's state is
STARTING it gets changed to PLAYING just before calling
DSOUND_MixOne(). It has been like that since
6e8c8c28ebbdc15292612b7b6657e44ce8c5040f (17 years ago), mere
months after the lead-in logic was introduced in
36e90546298e8c096926330dd75f4e373a415171.

I don't know whether that code does something useful. It was
introduced in a monster 390+ 310- commit whose commit message
doesn't say much, and the comment near the code only mentions
"more fluid pointer advancement", which I have no idea of
what it means. But the fact that it was effectively dead for
95% of its existence and nobody every complained seems to
suggest that we can get rid of it.