dsound: Use AUTOCONVERTPCM when initializing the render audio client.
Initializing the audio client without AUTOCONVERTPCM is expected
to fail when the requested sampling rate or channel count does not
match the mixing format. Such behavior is not currently modeled on
Wine, but it should be introduced, and since here we expect
automatic format conversion to work we add the flag.