mmdevapi: Move format validation in driver-independent code.
The formal validation of the audio format is the same for all
drivers, there is no point having duplicated and different
validation code in each driver.

The code I copied here is from the Pulse driver, which looks the
most complete. It probably can be improved, but so far I'm just
refactoring.

Not all driver-specific validation code is removed yet, because
the suggested format output must be handled before.