win32u: Update the window config before its state, as internal config.
When restoring a window, so that the window gets restored to the size
it currently has from the host point of view. Either the host already
has restored it to its original size, or it will shortly, but we should
avoid requesting config changes concurrently if possible.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=58804