winevulkan: Update to VK spec version 1.4.335.
We should support VK_EXT_present_timing at some point,
but it's not trivial to handle and I won't have time for it
before the code freeze (nor a linux driver to test).

For now, at least update the spec to ship the
VK_EXT_ray_tracing_invocation_reorder fix.