win32u: Use the window fullscreen exclusive rect for vulkan swapchain.
The present rect, if set, is used for creating the client surfaces. We
need to do the same to create vulkan swapchains, in case the WineD3D
vulkan renderer is being used.

This will also be needed to implement the exclusive fullscreen mode
mechanism in DXVK and fix the reported regression.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=58443